+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
   - Ignore object ids that are too far off
+  - Improve filter speed
 
 ## [0.4.0] - 2025-09-17